# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Mist Automation & Backup API
  description: |
    Comprehensive web application for automating Juniper Mist operations and managing configuration backups.
    
    ## Authentication
    This API uses JWT Bearer token authentication. To authenticate:
    1. Call `POST /api/v1/auth/login` with your credentials
    2. Use the returned `access_token` in the `Authorization` header as `Bearer <token>`
    
    ## First-Time Setup (Onboarding)
    When the system has no users, use `POST /api/v1/auth/onboard` to create the first admin user.
    
    ## Rate Limiting
    API requests are rate-limited per user. Please handle 429 responses appropriately.
    Login attempts are rate-limited to 5 attempts per 5 minutes per IP/email combination.
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
    description: API v1
...
security:
  - BearerAuth: []
paths:
...
  /webhooks/mist:
    post:
      tags:
        - Webhooks
      summary: Receive Mist webhook
      description: |
        Unified Mist webhook gateway. Receives all Mist webhooks, stores them as WebhookEvents,
        and routes internally to both the automation and backup modules.
        Called by Mist Cloud or Smee.io forwarder.
      operationId: receiveMistWebhook
      security: []
      parameters:
        - name: X-Mist-Signature
          in: header
          description: Mist webhook HMAC-SHA256 signature for verification
          schema:
            type: string
        - name: X-Forwarded-By
          in: header
          description: Set by internal Smee forwarder (value 'smee' bypasses signature check from localhost)
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: Mist webhook payload
      responses:
        '200':
          description: Webhook received and processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookReceivedResponse'
        '401':
          description: Missing or invalid webhook signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /webhooks/events:
    get:
      tags:
        - Webhooks
      summary: List webhook events
      description: List webhook events received from Mist
      operationId: listWebhookEvents
      parameters:
        - name: skip
          in: query
          description: Number of events to skip
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          description: Number of events to return
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: webhook_type
          in: query
          description: Filter by webhook type
          schema:
            type: string
        - name: processed
          in: query
          description: Filter by processed status
          schema:
            type: boolean
      responses:
        '200':
          description: List of webhook events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookListResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /webhooks/events/{event_id}:
    get:
      tags:
        - Webhooks
      summary: Get webhook event
      description: Get webhook event details by ID including full payload and headers
      operationId: getWebhookEvent
      parameters:
        - name: event_id
          in: path
          required: true
          description: Webhook event ID
          schema:
            type: string
      responses:
        '200':
          description: Webhook event details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookEventDetailResponse'
        '400':
          description: Invalid event ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /webhooks/events/{event_id}/replay:
    post:
      tags:
        - Webhooks
      summary: Replay webhook event
      description: Replay a webhook event through the workflow engine
      operationId: replayWebhookEvent
      parameters:
        - name: event_id
          in: path
          required: true
          description: Webhook event ID
          schema:
            type: string
      responses:
        '200':
          description: Replay queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusMessageResponse'
        '400':
          description: Invalid event ID format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /webhooks/stats:
    get:
      tags:
        - Webhooks
      summary: Get webhook statistics
      description: Get aggregated webhook volume statistics bucketed by time
      operationId: getWebhookStats
      parameters:
        - name: hours
          in: query
          description: Time range in hours (max 30 days)
          schema:
            type: integer
            minimum: 1
            maximum: 720
            default: 24
      responses:
        '200':
          description: Webhook statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookStatsResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /webhooks/smee/status:
    get:
      tags:
        - Webhooks
      summary: Get Smee status
      description: Get Smee.io webhook forwarder client status (admin only)
      operationId: getSmeeStatus
      responses:
        '200':
          description: Smee status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmeeStatusResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /webhooks/smee/start:
    post:
      tags:
        - Webhooks
      summary: Start Smee client
      description: |
        Start the Smee.io webhook forwarder (admin only).
        Accepts an optional smee_channel_url in the request body.
        The provided URL is persisted automatically.
      operationId: startSmeeClient
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmeeStartRequest'
      responses:
        '200':
          description: Smee started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmeeStatusResponse'
        '400':
          description: Smee.io channel URL not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /webhooks/smee/stop:
    post:
      tags:
        - Webhooks
      summary: Stop Smee client
      description: Stop the Smee.io webhook forwarder (admin only)
      operationId: stopSmeeClient
      responses:
        '200':
          description: Smee stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusMessageResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ═══════════════════════════════════════════════════════════════════════════
  # Backup Endpoints
  # ═══════════════════════════════════════════════════════════════════════════
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
    StatusMessageResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
        message:
          type: string

    # ─────────────────────────────────────────────────────────────────────────
    # Health
    # ─────────────────────────────────────────────────────────────────────────
...
    WebhookReceivedResponse:
      type: object
      required:
        - status
        - event_id
        - routed_to
        - message
      properties:
        status:
          type: string
          examples:
            - received
        event_id:
          type: string
          description: Stored event ID
        routed_to:
          type: array
          items:
            type: string
          description: Modules the webhook was routed to
        message:
          type: string
        backup_result:
          type:
            - object
            - 'null'
          additionalProperties: true
          description: Result from backup module if applicable
    WebhookEventResponse:
      type: object
      required:
        - id
        - webhook_type
        - webhook_id
        - processed
        - signature_valid
        - received_at
      properties:
        id:
          type: string
          description: Event ID
        webhook_type:
          type: string
          description: Webhook type
        webhook_topic:
          type:
            - string
            - 'null'
          description: Webhook topic
        webhook_id:
          type: string
          description: Unique webhook ID
        source_ip:
          type:
            - string
            - 'null'
          description: Source IP address
        site_id:
          type:
            - string
            - 'null'
          description: Mist site ID
        org_id:
          type:
            - string
            - 'null'
          description: Mist organization ID
        processed:
          type: boolean
          description: Whether processed
        matched_workflows:
          type: array
          items:
            type: string
          description: Matched workflow IDs
        executions_triggered:
          type: array
          items:
            type: string
          description: Triggered execution IDs
        signature_valid:
          type: boolean
          description: Signature validity
        routed_to:
          type: array
          items:
            type: string
          description: Modules routed to
        response_status:
          type: integer
          description: HTTP status code returned
        response_body:
          type: object
          additionalProperties: true
          description: Response body returned
        received_at:
          type: string
          format: date-time
          description: Receipt timestamp
        processed_at:
          type:
            - string
            - 'null'
          format: date-time
          description: Processing timestamp
    WebhookEventDetailResponse:
      allOf:
        - $ref: '#/components/schemas/WebhookEventResponse'
        - type: object
          required:
            - payload
          properties:
            payload:
              type: object
              additionalProperties: true
              description: Webhook payload
            headers:
              type: object
              additionalProperties:
                type: string
              description: HTTP headers
    WebhookListResponse:
      type: object
      required:
        - events
        - total
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/WebhookEventResponse'
          description: List of webhook events
        total:
          type: integer
          description: Total number of events
    WebhookStatsBucket:
      type: object
      required:
        - bucket
        - total
        - by_topic
      properties:
        bucket:
          type: string
          description: Time bucket label (ISO format)
        total:
          type: integer
          description: Total events in bucket
        by_topic:
          type: object
          additionalProperties:
            type: integer
          description: Event counts by webhook topic
    WebhookStatsResponse:
      type: object
      required:
        - buckets
        - granularity
        - hours
      properties:
        buckets:
          type: array
          items:
            $ref: '#/components/schemas/WebhookStatsBucket'
          description: Time-bucketed statistics
        granularity:
          type: string
          enum: [hourly, daily]
          description: Bucket granularity
        hours:
          type: integer
          description: Time range in hours

    # ─────────────────────────────────────────────────────────────────────────
    # Workflow Simulation & Variables
    # ─────────────────────────────────────────────────────────────────────────
...
    SmeeStatusResponse:
      type: object
      required:
        - running
      properties:
        running:
          type: boolean
          description: Whether Smee client is running
        channel_url:
          type:
            - string
            - 'null'
          description: Smee channel URL
        status:
          type: string
          description: Status message
    SmeeStartRequest:
      type: object
      properties:
        smee_channel_url:
          type: string
          description: Smee.io channel URL (optional, uses saved config if not provided)

    # ─────────────────────────────────────────────────────────────────────────
    # Admin
    # ─────────────────────────────────────────────────────────────────────────
...